+++ test/libdnf/transaction/test_rpm_package.cpp
@@ -89,7 +89,7 @@ void TransactionRpmPackageTest::test_save_load() {
 
     // get the written transaction
     auto ts_list = base2->get_transaction_history()->list_transactions({trans.get_id()});
-    CPPUNIT_ASSERT_EQUAL(1LU, ts_list.size());
+    CPPUNIT_ASSERT_EQUAL((size_t)1, ts_list.size());
 
     auto trans2 = ts_list[0];
